\name{print.hwtest}
\alias{print.hwtest}
\title{S3 Method for printing \code{hwtest} objects}
\usage{
\method{print}{hwtest}(x, ...)
}
\arguments{
  \item{x}{the results from a call to
  \code{\link{hw.test}}}

  \item{...}{other parameters passed to \code{print}. You
  can specify \code{detail} or \code{statName}}
}
\description{
Prints test results (\code{hwtest}) objects depending on
how much detail is provided.
}

