% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hwtest.R
\name{print.hwtest}
\alias{print.hwtest}
\title{S3 Method for printing \code{hwtest} objects}
\usage{
\method{print}{hwtest}(x, detail = NA, statName = NA, plotHisto = TRUE,
  ...)
}
\arguments{
\item{x}{the results from a call to \code{\link{hwx.test}}}

\item{detail}{0 for no print; 1 for P value only; 2 for all four P values; 3 to add data; 4 to add expected values}

\item{statName}{which statistic to use}

\item{plotHisto}{Indicate whether or not to plot the histogram. Only used if \code{\link{hwx.test}} was called with \code{histobins} set to a positive value.}

\item{\dots}{other parameters passed to \code{print}.}
}
\description{
Prints test results (\code{hwtest}) objects depending on how much detail is provided. If histogram data are present, \code{ggplot2} is used to draw the plot by calling \code{\link{makeHistogram}}
}

