% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.hwtest}
\alias{print.hwtest}
\title{S3 Method for printing \code{hwtest} objects}
\usage{
\method{print}{hwtest}(x, detail = NA, statName = NA, ...)
}
\arguments{
\item{x}{the results from a call to \code{\link{hwx.test}}}

\item{detail}{0 for no print; 1 for P value only; 2 for all four P values; 3 to add data; 4 to add expected values}

\item{statName}{which statistic to use}

\item{\dots}{other parameters passed to \code{print}.}
}
\description{
Prints test results (\code{hwtest}) objects depending on how much detail is provided.
}

